DAT-1024 refactor typesmanager
