+++ cmdb/exporter/exporter_base.py
@@ -26,7 +26,6 @@ import zipfile
 import openpyxl
 
 from cmdb.database.database_manager_mongo import DatabaseManagerMongo
-from cmdb.manager.type_manager import TypeManager
 
 from cmdb.utils import json_encoding
 from cmdb.utils.helpers import load_class
@@ -41,7 +40,6 @@ from cmdb.framework.cmdb_render import RenderResult
 LOGGER = logging.getLogger(__name__)
 
 database_manager = DatabaseManagerMongo(**SystemConfigReader().get_all_values_from_section('Database'))
-type_manager = TypeManager(database_manager=database_manager)
 
 # -------------------------------------------------------------------------------------------------------------------- #
 #                                                     ZipExportType                                                    #
