+++ cmdb/framework/datagerry_assistant/profile_base.py
@@ -19,7 +19,7 @@ This module is the base class for the profiles of DATAGERRY assistant
 import logging
 from flask import current_app
 
-from cmdb.manager.type_manager import TypeManager
+from cmdb.manager.types_manager import TypesManager
 
 from cmdb.framework import TypeModel
 from .profile_type_constructor import ProfileTypeConstructor
@@ -41,7 +41,7 @@ class ProfileBase:
         self.type_collection = TypeModel.COLLECTION
 
         with current_app.app_context():
-            self.type_manager = TypeManager(current_app.database_manager)
+            self.types_manager = TypesManager(current_app.database_manager)
             self.type_constructor = ProfileTypeConstructor(current_app.database_manager)
 
 # ------------------------------------------------- HELPER FUNCTIONS ------------------------------------------------- #
@@ -65,8 +65,8 @@ class ProfileBase:
             type_name_id (str): Key which should be used for the id of this type, like 'user_type_id'
             type_dict (dict): all the required data to create the type except the public_id
         """
-        type_dict['public_id'] = self.type_manager.get_new_id()
-        new_type_id: int = self.type_manager.insert(type_dict)
+        type_dict['public_id'] = self.types_manager.get_new_type_public_id()
+        new_type_id: int = self.types_manager.insert_type(type_dict)
 
         self.created_type_ids[type_name_key] = new_type_id
